Separate TypeModel.cs types into separate files
