+++ XmlSchemaClassGenerator/Models/DocumentationModel.cs
@@ -1,0 +1,7 @@
+﻿namespace XmlSchemaClassGenerator;
+
+public class DocumentationModel
+{
+    public string Language { get; set; }
+    public string Text { get; set; }
+}
